requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
el==0.0.4
elevenlabs==1.54.0
fastapi==0.115.11
ffmpeg==1.4
flet==0.27.6
flet-desktop==0.27.6
flet-web==0.27.6
frozenlist==1.5.0
gigachat==0.1.39.post1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.47
langchain-gigachat==0.3.8
langsmith==0.3.18
magic-filter==1.0.12
multidict==6.2.0
numpy==2.2.4
oauthlib==3.2.2
orjson==3.10.15
packaging==24.2
propcache==0.3.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.0.1
PyYAML==6.0.2
repath==0.9.0
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.46.1
tenacity==9.0.0
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wavio==0.0.9
websockets==15.0.1
yarl==1.18.3
zstandard==0.23.0
